withJSON/requirements.txt
#utilities
numpy==1.24.4
opencv-python-headless==4.11.0.86
onnxruntime==1.14.1
kafka-python==2.0.2
pytesseract==0.3.10



